+++ source/platforms/platform.ts
@@ -2,6 +2,7 @@ import { Env, CISource } from "../ci_source/ci_source"
 import { GitDSL } from "../dsl/GitDSL"
 import { GitHub } from "./GitHub"
 import { GitHubAPI } from "./github/GitHubAPI"
+import { LocalGit } from "./local-git"
 
 /** A type that represents the downloaded metadata about a code review session */
 export type Metadata = any
@@ -54,6 +55,10 @@ export interface Platform {
  * @returns {Platform} returns a platform if it can be supported
  */
 export function getPlatformForEnv(env: Env, source: CISource): Platform {
+  if (source.name === "local repo") {
+    return new LocalGit()
+  }
+
   const token = env["DANGER_GITHUB_API_TOKEN"]
   if (!token) {
     console.error("The DANGER_GITHUB_API_TOKEN environmental variable is missing")